+++ src/rmi/InterfaceServ.java
@@ -13,7 +13,7 @@ import servidor.Leilao;
 
 public interface InterfaceServ extends Remote  {
     public void chamar(String nome, InterfaceCli cliente)throws RemoteException;
-    public void cadastrarLeilao(String nome_produto, String descricao, float preco, float tempo, InterfaceCli cliente) throws RemoteException;
+    public void cadastrarLeilao(String nome_produto, String descricao, float preco, int tempo, InterfaceCli cliente) throws RemoteException;
     public ArrayList<Leilao> leiloes_ativos() throws RemoteException;
     public String registarLance(String codigo, float valor, InterfaceCli cliente) throws RemoteException;
     public void cancelarLeilao(String codigo) throws RemoteException;
